+++ iowmongotools/__init__.py
@@ -7,22 +7,3 @@ import logging
 from iowmongotools import app
 
 logger = logging.getLogger(__name__)
-
-
-class MongoCheck(app.App):
-
-    @property
-    def default_config(self):
-        config = super().default_config
-        config.update({
-            'clusters': (('or', 'sc', 'eu', 'jp'), 'Set of a designation of clusters'),
-            'upload_dir': ('upload', 'Directory for uploads')
-        })
-        return config
-
-
-class MongoCheckCli(MongoCheck, app.AppCli):
-
-    def run(self):
-        logger.debug('wefv')
-        return 0